requirements.txt
dj-database-url==0.4.2
Django==1.11
djangorestframework==3.6.2
mock==2.0.0
pbr==2.1.0
psycopg2==2.7.1
pytz==2017.2
requests==2.13.0
six==1.10.0
